Adds the Fleet of aliens
Added the row with the Fleet of Aliens for the game.
Determined How Many aliens Fit in a Row.
The screen width is stored in settings.screen_width, but we
need an empty margin on either side of the screen. We'll make this
margin the width of the one alien. Because we have two margins, the
available space fot aliens is the screen width minus two alien widths.
available_space_x = settings.screen_width - (2 * alien_width)

Determined also the number of aliens in the Fleet:
We used floor division (//), which divides two numbers and drops any
remainer, we will get an integer number of aliens.
number_aliens_x = available_space_x // (2 * alien_width).
Each alien is pushed to the right of the previous one.

Signed-off-by: Francisco Pedraza <pacopedraza87@gmail.com>